+++ src/str/charsplit_test.h
@@ -1,6 +1,6 @@
 #ifndef ALI__STR_CHARSPLIT_TEST_H
 #define ALI__STR_CHARSPLIT_TEST_H
 
-size_t test_ali_str_charsplit(void);
+int test_ali_str_charsplit(void);
 
 #endif
